\name{add-package}
\alias{add-package}
\docType{package}
\title{
\packageTitle{add}
}
\description{
\packageDescription{add}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{add}
\packageIndices{add}

The package is a minimal example including function \code{add()} which
is trivial.
}
\author{
\packageAuthor{add}

Maintainer: \packageMaintainer{add}
}
\keyword{ package }
\examples{
add(2,3)
}
