% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ald_audit.R
\name{ald_audit}
\alias{ald_audit}
\title{ALD Audit}
\usage{
ald_audit(
  file,
  outcome_variable = NULL,
  prediction_variable = NULL,
  ground_truth_variable = NULL,
  sensitive_attributes = NULL,
  notion_of_fairness,
  ranking_mechanism,
  adjust_method = "BH",
  data_name = NULL,
  dir = here::here(""),
  n_grp,
  ntree,
  alpha,
  seed = NULL
)
}
\arguments{
\item{file}{.csv file that holds the data to conduct the audit on, character(1)}

\item{outcome_variable}{name of the outcome variable in your data file as character(1) or NULL (by default). Only in case notion_of_fairness is 'statistical parity' do you require to give an outcome variable}

\item{prediction_variable}{name of the prediction variable in your data file as character(1) or NULL (by default). Only in case notion_of_fairness is 'equalised odds' do you require to give an prediction variable}

\item{ground_truth_variable}{name of the ground truth variable in your data file as character(1) or NULL (by default). Only in case notion_of_fairness is 'equalised odds' do you require to give an ground truth variable}

\item{sensitive_attributes}{names of the sensitive attribute variables in your data file as character string or NULL (by default) t use all variables that are not the outcome or prediction and ground truth variables}

\item{notion_of_fairness}{notion of fairness must be 'statistical parity' or 'equalized odds'}

\item{ranking_mechanism}{ranking mechanism must be 'confidence' or 'magnitude'}

\item{adjust_method}{adjust method must be 'bonferroni' or 'BH' (by default)}

\item{data_name}{name used in the file name of the resulted report character(1)}

\item{dir}{directory to which to write the report}

\item{n_grp}{maximum number of groups in the report, numeric(1)}

\item{ntree}{number of trees to model in partykit::cforest, numeric(1)}

\item{alpha}{alpha argument passed to partykit::cforest, numeric(1)}

\item{seed}{random seed number can be used for reproducibility of results, numeric(1)}
}
\value{
pdf report
}
\description{
The ALD audit:
\itemize{
  \item is performed on a dataset of your choice, the dataset must be provided as a .csv file
  \item requires notion of fairness to be set to 'statistical parity' or 'equalized odds'
  \item in case of 'statistical parity' you must set the outcome_variable argument to the name of the outcome variable in your dataset
  \item for 'equalized odds' you must set the prediction_variable and ground_truth_variable arguments to the names of the prediction and ground truth variables in your dataset
  \item by default all other variables (not outcome, prediction, ground truth) in you dataset will be used as sensitive attributes in the audit. You can use the sensitive_attributes argument to specifically set the sensitive attributes to a subset of your dataset varaiables
  \item requires a ranking mechanism which must be 'confidence' or 'magnitude'
  \item requires a maximum number of groups in the report (n_grp)
  \item requires a number of trees to model in partykit::cforest (ntree)
  \item requires a alpha argument passed to partykit::cforest (alpha)
  \item optionally takes a adjustment method ("BH" Benjamini-Hochberg by default) to pass to stats::p.adjust (adjust_method)
  \item optionally takes a random seed number that can be used for reproducibility of results
  \item writes a report to the directory that you set with the dir argument, with data_name argument used in the name
}
}
