% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_unique_leaves.R
\name{analyse_unique_leaves}
\alias{analyse_unique_leaves}
\title{Analyse unique leaves}
\usage{
analyse_unique_leaves(partitioning, dataset, psi_metric, ntree)
}
\arguments{
\item{partitioning}{result from partykit::cforest()}

\item{dataset}{data frame that the cforest was trained on}

\item{ntree}{numeric(1) number of trees}
}
\description{
Analyse unique leaves
}
