% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{apply_split}
\alias{apply_split}
\title{Apply split on data}
\usage{
apply_split(cnode, data, side)
}
\arguments{
\item{cnode}{node}

\item{data}{data frame to split}

\item{side}{numeric(1) side 1 or 2}
}
\description{
Apply split on data
}
