% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{audit_report_table}
\alias{audit_report_table}
\title{Create the table for the audit report}
\usage{
audit_report_table(results_df, n_grp, dataset)
}
\arguments{
\item{results_df}{data frame, results from the analysis}

\item{n_grp}{numeric(1), number of groups}

\item{dataset}{data frame that the cforest was trained on}
}
\description{
Creates a flextable
}
