% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{audit_report_trees}
\alias{audit_report_trees}
\title{Create audit report trees}
\usage{
audit_report_trees(results_df, partitioning, dataset, psi_metric, sen_attr)
}
\arguments{
\item{results_df}{data frame, results from the analysis}

\item{partitioning}{result from partykit::cforest()}

\item{dataset}{data frame that the cforest was trained on}

\item{psi_metric}{character(1)}

\item{sen_attr}{character()}
}
\description{
Get the n_grp first number of trees and prune if necessary. Create the visualisations.
}
\details{
Return the visualisations and pruning message
}
