% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{clean_condition_str}
\alias{clean_condition_str}
\title{Clean a condition string}
\usage{
clean_condition_str(raw_str)
}
\arguments{
\item{raw_str}{character(1)}
}
\description{
Clean a condition string
}
