% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{compute_abs_odds}
\alias{compute_abs_odds}
\title{Compute average absolute odds difference for given dataset}
\usage{
compute_abs_odds(rest_error_type, subset_error_type)
}
\arguments{
\item{rest_error_type}{}

\item{subset_error_type}{}
}
\description{
Compute average absolute odds difference for given dataset
}
