% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_unique_leaves.R
\name{compute_magnit_and_confid}
\alias{compute_magnit_and_confid}
\title{compute magnitude and (uncorrected) confidence for given leaf (subgroup)}
\usage{
compute_magnit_and_confid(.data, psi_metric)
}
\arguments{
\item{.data}{data frame with a leaf_indicator_ column and outcome (for statistical parity) or error_type (for equalized odds)}

\item{psi_metric}{character(1)}
}
\description{
compute magnitude and (uncorrected) confidence for given leaf (subgroup)
}
