% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{compute_stat_par}
\alias{compute_stat_par}
\title{Compute statistical parity difference directly for two subsets}
\usage{
compute_stat_par(rest_outcome, subset_outcome, x = 1)
}
\arguments{
\item{rest_outcome}{}

\item{subset_outcome}{}

\item{x}{}
}
\description{
Compute statistical parity difference directly for two subsets
}
