% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{export_audit_report}
\alias{export_audit_report}
\title{Create and save an audit report}
\usage{
export_audit_report(
  results_df,
  n_grp,
  psi_metric,
  ranking,
  adj_method,
  sen_attr,
  partitioning,
  dataset,
  data_name,
  dir
)
}
\arguments{
\item{results_df}{data frame, results from the analysis}

\item{n_grp}{numeric(1),number of groups}

\item{psi_metric}{character(1)}

\item{ranking}{character(1)}

\item{adj_method}{character(1)}

\item{sen_attr}{character()}

\item{partitioning}{result from partykit::cforest()}

\item{dataset}{data frame that the cforest was trained on}

\item{data_name}{character(1) name of the dataset}

\item{dir}{character(1) name of directory to write to}
}
\description{
Audit report contains a title, description, table and 3 tree plots.
}
\details{
It writes the file to <dir>/audit_report_<data_name>.pdf
}
