% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{get_node_depths}
\alias{get_node_depths}
\title{Get depths for nodes of trees}
\usage{
get_node_depths(tree)
}
\arguments{
\item{tree}{result from partykit::gettree()}
}
\description{
https://stackoverflow.com/questions/35265552/depth-of-a-node-in-partykit
}
