% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_unique_leaves.R
\name{get_total_n_leaves}
\alias{get_total_n_leaves}
\title{Get total number of leaves}
\usage{
get_total_n_leaves(partitioning, ntree)
}
\arguments{
\item{partitioning}{the result of partykit::cforest}

\item{ntree}{numeric(1) number of trees}
}
\description{
Get total number of leaves in all trees of causal forest
}
