% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(
  file,
  outcome,
  prediction,
  ground_truth,
  sen_attr,
  psi_metric,
  reserved_colnames
)
}
\arguments{
\item{file}{character(1)}

\item{outcome}{character(1)}

\item{prediction}{character(1)}

\item{ground_truth}{character(1)}

\item{sen_attr}{character()}

\item{psi_metric}{character(1)}

\item{reserved_colnames}{character()}
}
\description{
- Read csv file 
- In case of statistical parity 
 + check and set the outcome variable
- In case of equalized odds
 + check the prediction and ground_truth
 + set the error_type and outcome variables
- check and set the sensitive attribute variables
- select the outcome and sensitive attributes as columns, plus the error_type in case of equalized odds
- set character variables to factors
}
