% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess_results}
\alias{postprocess_results}
\title{Postprocess results}
\usage{
postprocess_results(results_df, ranking)
}
\arguments{
\item{results_df}{data frame, results from analyse_unique_leaves()}

\item{ranking}{character(1)}
}
\description{
- Clean each group_condition
- Remove duplicate conditions
- Adjust disparity confidence
- Rank results by selected ranking
}
