% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{rank_results}
\alias{rank_results}
\title{Rank results according to either magnitude or confidence}
\usage{
rank_results(results_df, ranking)
}
\arguments{
\item{results_df}{data frame}

\item{ranking}{character(1) 'confidence' or 'magnitude'}
}
\description{
Rank results according to either magnitude or confidence
}
