% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{rec_nodes}
\alias{rec_nodes}
\title{Recursively compute metrics for all nodes (not only leaves)}
\usage{
rec_nodes(cnode, data, subset, psi_metric, first = FALSE)
}
\arguments{
\item{cnode}{node}

\item{data}{full dataset}

\item{subset}{subset of node}

\item{psi_metric}{character(1)}

\item{first}{logical(1) Initial computation of recursive flow}
}
\value{
metrics <- rec_nodes(cnode, dataset, dataset, psi_metric = psi_metric, first = TRUE)
}
\description{
Recursively compute metrics for all nodes (not only leaves)
}
