% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{visualize_tree}
\alias{visualize_tree}
\title{Visualize given tree}
\usage{
visualize_tree(results_df, t, ct, dataset, psi_metric)
}
\arguments{
\item{results_df}{data frame, results from the analysis}

\item{t}{numeric(1) tree number}

\item{ct}{partykit tree}

\item{dataset}{data frame that the cforest was trained on}

\item{psi_metric}{character(1)}
}
\description{
- Recursively analyse nodes
- plot
}
