\name{add}
\alias{add}
\alias{subtract}
\title{Adds two numbers}
\description{Trivial function that adds two numbers.}
\usage{
add(x, y)
subtract(x, y)
}
\arguments{
  \item{x,y}{Numbers to be added}
}
\details{
Given two numbers, return their sum.
}
\value{Another number}
\author{Robin K. S. Hankin}
\note{
Function \code{add()} takes vectors and matrices}
\examples{
add(3,4)   # should be 7
}
